+++ client/src/components/common/__test__/Header.test.tsx
@@ -1,18 +1,30 @@
 import { render, screen } from '@testing-library/react';
 import Header from '../Header';
-import { AppRouterContextProviderMock } from '../../../../tests/AppRouterContextProviderMock';
 import userEvent from '@testing-library/user-event';
+import {
+  AppRouterContext,
+  AppRouterInstance,
+} from 'next/dist/shared/lib/app-router-context';
+
+// mock route because we need navigation context
+const mockedRouter: AppRouterInstance = {
+  back: jest.fn(),
+  forward: jest.fn(),
+  push: jest.fn(),
+  replace: jest.fn(),
+  refresh: jest.fn(),
+  prefetch: jest.fn(),
+};
 
 describe('Header component', () => {
   it('redirects to /home when logo is clicked', async () => {
-    const push = jest.fn();
     render(
-      <AppRouterContextProviderMock router={{ push }}>
+      <AppRouterContext.Provider value={mockedRouter}>
         <Header user={{ name: 'John', email: 'test1@gmail.com', id: '1' }} />
-      </AppRouterContextProviderMock>
+      </AppRouterContext.Provider>
     );
-    const logo = screen.getByText('finder.com');
-    await userEvent.click(logo);
-    expect(push).toHaveBeenCalledWith('/home');
+    const linkElement = screen.getByTestId('finder-logo');
+    await userEvent.click(linkElement);
+    expect(linkElement).toHaveAttribute('href', '/home');
   });
 });
